more chuncks from infinity sd2 ^&^ fixes ( sql work to follow when ever i get the DB up and running 8p )

+++ scripts/northrend/icecrown_citadel/frozen_halls/halls_of_reflection/halls_of_reflection.cpp
@@ -952,7 +952,7 @@ struct MANGOS_DLL_DECL npc_jaina_and_sylvana_HRextroAI : public npc_escortAI
          }
          else
          {
-             if (HoldTimer <= diff)
+             if (HoldTimer <= (int)diff)
                  HoldTimer = 0;
              else
                  HoldTimer -= diff;
